Reorganization to match code style.
